backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

transformers
torch
sentence-transformers

langchain
faiss-cpu

python-dotenv

pytest
pytest-asyncio

chromadb